+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.3.0
 
 - update RxDart to `0.21.0`;
+- rename `onRunning` to `running`
 
 ## 0.2.0